refactor: move IconButton to SidebarButton
