+++ app/src/components/Sidebar/MobileSidebarOpenControls.tsx
@@ -1,6 +1,6 @@
 import { faBars } from '@fortawesome/free-solid-svg-icons';
-import IconButton from '@components/IconButton';
 import Logo from '@components/Logo';
+import SidebarButton from './SidebarButton';
 
 type MobileSidebarOpenControlsProps = {
   onOpenSidebar: () => void;
@@ -10,7 +10,7 @@ const MobileSidebarOpenControls = ({
   onOpenSidebar,
 }: MobileSidebarOpenControlsProps) => (
   <div className="flex items-center justify-between pb-4 lg:hidden">
-    <IconButton label="open menu" icon={faBars} onClick={onOpenSidebar} />
+    <SidebarButton label="open menu" icon={faBars} onClick={onOpenSidebar} />
     <Logo variant="primary" />
   </div>
 );
